app/requirements.txt
Flask-RESTful==0.3.7
Flask==1.1.1
gunicorn==20.0.4
python-dotenv==0.19.0
flask-apispec
coverage